+++ spec/lib/tasks/users/create_spec.rb
@@ -1,24 +1,36 @@
 require "rails_helper"
 
 RSpec.describe "rake users:create", type: :task do
+  let(:email) { "john.doe@education.gov.uk" }
+  let(:first_name) { "John" }
+  let(:last_name) { "Doe" }
+
+  subject { Rake::Task["users:create"] }
+
+  after { subject.reenable }
+
+  describe "missing arguments" do
+    %w[email first_name last_name].each do |attribute|
+      context "when #{attribute} is not provided" do
+        let(:"#{attribute}") { nil }
+
+        it "aborts the task" do
+          expect { subject.invoke(email, first_name, last_name) }.to raise_error(SystemExit, I18n.t("tasks.users.create.error"))
+        end
+      end
+    end
+  end
+
   it "creates a user in the service" do
-    ClimateControl.modify(EMAIL_ADDRESS: "name@example.com") do
-      expect { Rake::Task["users:create"].execute }.to change { User.count }.by(1)
-      expect(User.last.email).to eq "name@example.com"
-    end
+    expect { subject.invoke(email, first_name, last_name) }.to change { User.count }.by(1)
+    expect(User.last).to have_attributes(first_name:, last_name:, email:)
   end
 
-  it "does not create duplicates" do
-    User.create(email: "name@example.com")
+  context "when the user already exists" do
+    let!(:user) { create(:user, email: email) }
 
-    ClimateControl.modify(EMAIL_ADDRESS: "name@example.com") do
-      expect { Rake::Task["users:create"].execute }
-        .to raise_error(ActiveRecord::RecordNotUnique)
+    it "does not create a duplicate" do
+      expect { subject.invoke(email, first_name, last_name) }.to raise_error(ActiveRecord::RecordNotUnique)
     end
-  end
-
-  it "errors if no email address is supplied" do
-    expect { Rake::Task["users:create"].invoke }
-      .to raise_error(SystemExit, I18n.t("tasks.users.create.error"))
   end
 end