+++ spec/models/user_spec.rb
@@ -51,31 +51,23 @@ RSpec.describe User do
     end
   end
 
+  describe "Validations" do
+    describe "#first_name" do
+      it { is_expected.to validate_presence_of(:first_name) }
+    end
+
+    describe "#last_name" do
+      it { is_expected.to validate_presence_of(:first_name) }
+    end
+  end
+
   describe "#full_name" do
+    let(:user) { build(:user) }
+
     subject { user.full_name }
 
-    context "when first name is nil" do
-      let(:user) { build(:user, first_name: nil) }
-
-      it "returns nil" do
-        expect(subject).to be_nil
-      end
-    end
-
-    context "when last name is nil" do
-      let(:user) { build(:user, last_name: nil) }
-
-      it "returns nil" do
-        expect(subject).to be_nil
-      end
-    end
-
-    context "when first name and last name are present" do
-      let(:user) { build(:user) }
-
-      it "returns full name" do
-        expect(subject).to eq "John Doe"
-      end
+    it "returns full name" do
+      expect(subject).to eq "John Doe"
     end
   end
 end